+++ src/connectedmachine/azext_connectedmachine/aaz/latest/connectedmachine/extension/_list.py
@@ -15,16 +15,16 @@ from azure.cli.core.aaz import *
     "connectedmachine extension list",
 )
 class List(AAZCommand):
-    """The operation to get all extensions of a Non-Azure machine.
+    """Get all extensions of a Non-Azure machine.
 
     :example: Sample command for extension list
         az connectedmachine extension list --machine-name myMachine --resource-group myResourceGroup
     """
 
     _aaz_info = {
-        "version": "2024-03-31-preview",
+        "version": "2024-05-20-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.hybridcompute/machines/{}/extensions", "2024-03-31-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.hybridcompute/machines/{}/extensions", "2024-05-20-preview"],
         ]
     }
 
@@ -133,7 +133,7 @@ class List(AAZCommand):
                     "$expand", self.ctx.args.expand,
                 ),
                 **self.serialize_query_param(
-                    "api-version", "2024-03-31-preview",
+                    "api-version", "2024-05-20-preview",
                     required=True,
                 ),
             }